requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-dotenv==1.0.0
openai>=1.0.0
pydantic==2.5.0
databases[sqlite]==0.9.0
aiosqlite==0.19.0
httpx>=0.25.0
google-cloud-documentai>=2.20.0
google-api-core>=2.11.0
google-generativeai>=0.3.0
pillow>=10.0.0
psycopg2-binary==2.9.9

